passed compiling

+++ algorithms/solvehfvc.cpp
@@ -9,15 +9,15 @@
 
 using std::cout;
 using std::endl;
-using Eigen::VectorXf;
-using Eigen::MatrixXf;
+using Eigen::VectorXd;
+using Eigen::MatrixXd;
 
 
-bool solvehfvc(const MatrixXf &N_ALL,
-  const MatrixXf &G, const VectorXf &b_G,
-  const VectorXf &F,
-  const MatrixXf &Aeq, const VectorXf &beq,
-  const MatrixXf &A, const VectorXf &b_A,
+bool solvehfvc(const MatrixXd &N_ALL,
+  const MatrixXd &G, const VectorXd &b_G,
+  const VectorXd &F,
+  const MatrixXd &Aeq, const VectorXd &beq,
+  const MatrixXd &A, const VectorXd &b_A,
   const int kDimActualized, const int kDimUnActualized,
   const int kDimSlidingFriction, const int kDimLambda,
   const int kNumSeeds,
@@ -41,26 +41,26 @@ bool solvehfvc(const MatrixXf &N_ALL,
   cout << "Begin solving for velocity commands" << endl;
   cout << "  [1] Determine Possible Dimension of control" << endl;
 
-  MatrixXf N = N_ALL.topRows(kDimLambda);
-  MatrixXf NG(N.rows()+G.rows(), N.cols());
+  MatrixXd N = N_ALL.topRows(kDimLambda);
+  MatrixXd NG(N.rows()+G.rows(), N.cols());
   NG << N, G;
 
   // matrix decomposition
-  Eigen::FullPivLU<MatrixXf> lu_decomp_N(N);
+  Eigen::FullPivLU<MatrixXd> lu_decomp_N(N);
   int rank_N = lu_decomp_N.rank();
-  MatrixXf basis_N = lu_decomp_N.kernel(); // columns of basis_N forms a basis
+  MatrixXd basis_N = lu_decomp_N.kernel(); // columns of basis_N forms a basis
                                            // of the null-space of N
-  Eigen::FullPivLU<MatrixXf> lu_decomp_NG(NG);
+  Eigen::FullPivLU<MatrixXd> lu_decomp_NG(NG);
   int rank_NG = lu_decomp_NG.rank();
-  MatrixXf basis_NG = lu_decomp_NG.kernel(); // columns of basis_NG forms a basis
+  MatrixXd basis_NG = lu_decomp_NG.kernel(); // columns of basis_NG forms a basis
                                            // of the null-space of NG
-  MatrixXf C_c(basis_NG.cols()+kDimUnActualized, basis_NG.rows());
+  MatrixXd C_c(basis_NG.cols()+kDimUnActualized, basis_NG.rows());
   C_c << basis_NG.transpose(),
-      MatrixXf::Identity(kDimUnActualized,kDimUnActualized),
-      MatrixXf::Zero(kDimUnActualized,kDimActualized);
-  Eigen::FullPivLU<MatrixXf>
+      MatrixXd::Identity(kDimUnActualized,kDimUnActualized),
+      MatrixXd::Zero(kDimUnActualized,kDimActualized);
+  Eigen::FullPivLU<MatrixXd>
       lu_decomp_C_c(C_c);
-  MatrixXf basis_c = lu_decomp_C_c.kernel(); // columns of basis_C_c forms a basis
+  MatrixXd basis_c = lu_decomp_C_c.kernel(); // columns of basis_C_c forms a basis
                                            // of the null-space of C_c
 
   int n_av = rank_NG - rank_N;
@@ -69,36 +69,36 @@ bool solvehfvc(const MatrixXf &N_ALL,
   cout << "   n_av = " << n_av << ", n_af = " << n_af << endl;
   assert(rank_N + kDimActualized >= kDimGeneralized);
 
-  MatrixXf R_a(kDimActualized, kDimActualized);
-  MatrixXf T(kDimGeneralized, kDimGeneralized);
-  VectorXf w_av;
+  MatrixXd R_a(kDimActualized, kDimActualized);
+  MatrixXd T(kDimGeneralized, kDimGeneralized);
+  VectorXd w_av;
   if (n_av == 0) {
     cout << "  [2] No feasible velocity control can satisfy the goal" << endl;
-    R_a = MatrixXf::Identity(kDimActualized, kDimActualized);
-    T = MatrixXf::Identity(kDimGeneralized, kDimGeneralized);
-    w_av = VectorXf(0);
+    R_a = MatrixXd::Identity(kDimActualized, kDimActualized);
+    T = MatrixXd::Identity(kDimGeneralized, kDimGeneralized);
+    w_av = VectorXd(0);
   } else {
     cout << "  [2] Solving for Directions by PGD" << endl;
 
     int NIter   = 50;
     int n_c     = rank_NG - kDimUnActualized;
-    MatrixXf BB = basis_c.transpose()*basis_c;
-    MatrixXf NN = basis_N*basis_N.transpose();
+    MatrixXd BB = basis_c.transpose()*basis_c;
+    MatrixXd NN = basis_N*basis_N.transpose();
 
-    std::vector<MatrixXf> k_all;
+    std::vector<MatrixXd> k_all;
     float cost_all[kNumSeeds] = {0};
     for (int seed = 0; seed < kNumSeeds; ++seed)  {
-      MatrixXf k  = MatrixXf::Random(n_c, n_av); // initial solution
-      MatrixXf bck = basis_c*k;
+      MatrixXd k  = MatrixXd::Random(n_c, n_av); // initial solution
+      MatrixXd bck = basis_c*k;
       for (int i = 0; i < bck.cols(); i++) {
           float bck_col_norm = bck.col(i).norm();
           k.col(i) /= bck_col_norm;
       }
-      MatrixXf g(n_c, n_av);
+      MatrixXd g(n_c, n_av);
       float costs = 0;
       for (int iter = 0; iter < NIter; ++iter) {
         // compute gradient
-        g = MatrixXf::Zero(n_c, n_av);
+        g = MatrixXd::Zero(n_c, n_av);
         costs = 0;
         for (int i = 0; i < n_av; ++i) {
           for (int j = 0; j < n_av; ++j) {
@@ -126,29 +126,29 @@ bool solvehfvc(const MatrixXf &N_ALL,
     const int SizeFloat = sizeof(cost_all) / sizeof(float);
     float *cost_best    = std::min_element(cost_all, cost_all + kNumSeeds);
     int min_id          = std::distance(cost_all, cost_best);
-    MatrixXf k_best     = k_all[min_id];
-    MatrixXf C_best     = (basis_c*k_best).transpose();
+    MatrixXd k_best     = k_all[min_id];
+    MatrixXd C_best     = (basis_c*k_best).transpose();
 
     // R_a = [null(C_best(:, kDimUnActualized+1:end))';
     //         C_best(:, kDimUnActualized+1:end)];
-    MatrixXf C_best_actualized = C_best.rightCols(kDimActualized);
-    Eigen::FullPivLU<MatrixXf> lu_decomp_C_best_actualized(C_best_actualized);
-    MatrixXf basis_C_best_actualized =
+    MatrixXd C_best_actualized = C_best.rightCols(kDimActualized);
+    Eigen::FullPivLU<MatrixXd> lu_decomp_C_best_actualized(C_best_actualized);
+    MatrixXd basis_C_best_actualized =
         lu_decomp_C_best_actualized.kernel(); // columns of basis_NG forms a basis
                   // of the null-space of C_best_actualized
-    R_a = MatrixXf::Zero(kDimActualized, kDimActualized);
+    R_a = MatrixXd::Zero(kDimActualized, kDimActualized);
     R_a << basis_C_best_actualized.transpose(), C_best_actualized;
-    T = MatrixXf::Zero(kDimGeneralized, kDimGeneralized);
+    T = MatrixXd::Zero(kDimGeneralized, kDimGeneralized);
     T.topLeftCorner(kDimUnActualized, kDimUnActualized) =
-        MatrixXf::Identity(kDimUnActualized, kDimUnActualized);
+        MatrixXd::Identity(kDimUnActualized, kDimUnActualized);
     T.bottomRightCorner(kDimActualized,kDimActualized) = R_a;
 
     // b_NG = [zeros(size(N, 1), 1); b_G];
-    VectorXf b_NG = VectorXf::Zero(N.rows() + b_G.rows());
+    VectorXd b_NG = VectorXd::Zero(N.rows() + b_G.rows());
     b_NG.tail(b_G.rows()) = b_G;
 
     // v_star = NG\b_NG;
-    VectorXf v_star = NG.fullPivLu().solve(b_NG);
+    VectorXd v_star = NG.fullPivLu().solve(b_NG);
     // cout << "C_best: " << C_best.rows() << ", " << C_best.cols() << endl;
     // cout << C_best;
 
@@ -158,35 +158,35 @@ bool solvehfvc(const MatrixXf &N_ALL,
   cout << "Begin Solving for force commands." << endl;
   // unactuated dimensions
   // H = [eye(kDimUnActualized), zeros(kDimUnActualized, kDimActualized)];
-  MatrixXf H = MatrixXf::Zero(kDimUnActualized, kDimGeneralized);
-  H.leftCols(kDimUnActualized) = MatrixXf::Identity(kDimUnActualized,
+  MatrixXd H = MatrixXd::Zero(kDimUnActualized, kDimGeneralized);
+  H.leftCols(kDimUnActualized) = MatrixXd::Identity(kDimUnActualized,
       kDimUnActualized);
   // Newton's laws
-  MatrixXf T_inv = T.inverse();
-  MatrixXf M_newton(kDimUnActualized+kDimGeneralized,
+  MatrixXd T_inv = T.inverse();
+  MatrixXd M_newton(kDimUnActualized+kDimGeneralized,
       kDimGeneralized + kDimContactForce);
   if (kDimSlidingFriction > 0) {
     M_newton.resize(kDimUnActualized+kDimGeneralized+Aeq.rows(),
       kDimGeneralized + kDimContactForce);
-    M_newton << MatrixXf::Zero(kDimUnActualized, kDimContactForce), H*T_inv,
+    M_newton << MatrixXd::Zero(kDimUnActualized, kDimContactForce), H*T_inv,
         T*N_ALL.transpose(),
-        MatrixXf::Identity(kDimGeneralized, kDimGeneralized), Aeq;
+        MatrixXd::Identity(kDimGeneralized, kDimGeneralized), Aeq;
   } else {
-    M_newton << MatrixXf::Zero(kDimUnActualized, kDimContactForce), H*T_inv,
+    M_newton << MatrixXd::Zero(kDimUnActualized, kDimContactForce), H*T_inv,
         T*N_ALL.transpose(),
-        MatrixXf::Identity(kDimGeneralized, kDimGeneralized);
+        MatrixXd::Identity(kDimGeneralized, kDimGeneralized);
   }
-  VectorXf b_newton(M_newton.rows());
+  VectorXd b_newton(M_newton.rows());
   if (kDimSlidingFriction > 0) {
-    b_newton << VectorXf::Zero(H.rows()), -T*F, beq;
+    b_newton << VectorXd::Zero(H.rows()), -T*F, beq;
   } else {
-    b_newton << VectorXf::Zero(H.rows()), -T*F;
+    b_newton << VectorXd::Zero(H.rows()), -T*F;
   }
 
-  MatrixXf M_free(M_newton.rows(), M_newton.cols() - n_af);
+  MatrixXd M_free(M_newton.rows(), M_newton.cols() - n_af);
   M_free << M_newton.leftCols(kDimContactForce+kDimUnActualized),
       M_newton.rightCols(n_av);
-  MatrixXf M_eta_af = M_newton.middleCols(kDimContactForce+kDimUnActualized,
+  MatrixXd M_eta_af = M_newton.middleCols(kDimContactForce+kDimUnActualized,
       n_af);
 
   // prepare the QP
@@ -204,7 +204,7 @@ bool solvehfvc(const MatrixXf &N_ALL,
   Eigen::MatrixXd G0 = Gdiag.asDiagonal();
   Eigen::VectorXd g0 = Eigen::VectorXd::Zero(G0.rows());
   // A_temp = [A(:, 1:kDimContactForce), A(:, kDimContactForce+1:end)*T_inv];
-  MatrixXf A_temp(A.rows(), A.cols());
+  MatrixXd A_temp(A.rows(), A.cols());
   A_temp << A.leftCols(kDimContactForce), A.rightCols(kDimGeneralized)*T_inv;
 
   // 0.5 x'Qx + f'x
@@ -213,22 +213,22 @@ bool solvehfvc(const MatrixXf &N_ALL,
   // A_eta_af = A_temp(:, kDimContactForce+kDimUnActualized+1:kDimContactForce+kDimUnActualized+n_af);
   // A_eta_av = A_temp(:, kDimContactForce+kDimUnActualized+n_af+1:end);
   // qp.A = [A_lambda_eta_u A_eta_av zeros(size(A, 1), n_dual_free) A_eta_af];
-  MatrixXf qpA(A.rows(), A.cols() + n_dual_free);
+  MatrixXd qpA(A.rows(), A.cols() + n_dual_free);
   qpA << A_temp.leftCols(kDimContactForce+kDimUnActualized),
-      A_temp.rightCols(n_av), MatrixXf::Zero(A.rows(), n_dual_free),
+      A_temp.rightCols(n_av), MatrixXd::Zero(A.rows(), n_dual_free),
       A_temp.middleCols(kDimContactForce+kDimUnActualized, n_af);
-  VectorXf qpb = b_A;
+  VectorXd qpb = b_A;
 
   // Aeq x = beq
   // qpAeq = [2*eye(n_free), M_free', zeros(n_free, n_af);
   //           M_free, zeros(size(M_free, 1)), M_eta_af];
   // qpbeq = [zeros(n_free, 1); b_newton];
-  MatrixXf qpAeq(n_free+M_free.rows(), A.cols() + n_dual_free);
-  qpAeq << 2*MatrixXf::Identity(n_free,n_free), M_free.transpose(),
-      MatrixXf::Zero(n_free, n_af), M_free,
-      MatrixXf::Zero(M_free.rows(), M_free.rows()), M_eta_af;
-  VectorXf qpbeq(n_free+b_newton.rows());
-  qpbeq << VectorXf::Zero(n_free), b_newton;
+  MatrixXd qpAeq(n_free+M_free.rows(), A.cols() + n_dual_free);
+  qpAeq << 2*MatrixXd::Identity(n_free,n_free), M_free.transpose(),
+      MatrixXd::Zero(n_free, n_af), M_free,
+      MatrixXd::Zero(M_free.rows(), M_free.rows()), M_eta_af;
+  VectorXd qpbeq(n_free+b_newton.rows());
+  qpbeq << VectorXd::Zero(n_free), b_newton;
 
   Eigen::VectorXd x = Eigen::VectorXd::Random(g0.rows());
   double cost = solve_quadprog(G0, g0,
@@ -239,12 +239,12 @@ bool solvehfvc(const MatrixXf &N_ALL,
       "]");
   cout << "  QP Solved. cost = " << cost << endl;
   // eta_af = x(n_free + n_dual_free + 1:end);
-  VectorXf eta_af = x.segment(n_free+n_dual_free, n_af).cast<float>();
+  VectorXd eta_af = x.segment(n_free+n_dual_free, n_af);
 
-  cout << "   Equality violation: " << (qpbeq - qpAeq*x.cast<float>()).norm()
+  cout << "   Equality violation: " << (qpbeq - qpAeq*x).norm()
       << endl;
 
-  VectorXf b_Ax = qpA*x.cast<float>() - qpb;
+  VectorXd b_Ax = qpA*x - qpb;
   float inequality_violation = 0;
   if (b_Ax.maxCoeff() > 0 ) inequality_violation = b_Ax.maxCoeff();
   cout << "   Inequality violation: " << inequality_violation
